% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-ext.R
\name{new_md_extension}
\alias{new_md_extension}
\title{MD extension}
\usage{
new_md_extension(content = character(), template = NULL, ...)
}
\arguments{
\item{template}{Path to a PowerPoint template.}

\item{...}{Forwarded to \code{\link[blockr.dock:extension]{blockr.dock::new_dock_extension()}}}

\item{document}{Initial document}
}
\description{
A markdown extension adds a document builder.
}
