% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-ui.R
\name{get_default_template}
\alias{get_default_template}
\title{Get default template selection}
\usage{
get_default_template()
}
\value{
Character string of default template
}
\description{
Internal helper to get the default template selection
}
\keyword{internal}
