% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-ui.R
\name{get_template_choices}
\alias{get_template_choices}
\title{Get template choices for UI}
\usage{
get_template_choices()
}
\value{
Named vector of template choices
}
\description{
Internal helper to get template choices based on available packages
}
\keyword{internal}
