% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_blockr_templates}
\alias{list_blockr_templates}
\title{List available PowerPoint templates}
\usage{
list_blockr_templates()
}
\value{
Character vector of available template names
}
\description{
List available PowerPoint templates
}
\examples{
list_blockr_templates()

}
