% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-board.R, R/md-render.R, R/md-server.R,
%   R/md-ui.R
\name{new_md_board}
\alias{new_md_board}
\alias{md_render}
\alias{md_render.file}
\alias{md_render.ggplot}
\alias{md_render.recordedplot}
\alias{md_render.evaluate_evaluation}
\alias{md_render.data.frame}
\alias{md_render.gt_tbl}
\alias{md_render.flextable}
\alias{md_server}
\alias{gen_md_ui}
\title{Markdown board}
\usage{
new_md_board(..., document = character())

md_render(x, value, dir = tempdir(), ...)

\method{md_render}{file}(x, value, dir = tempdir(), ...)

\method{md_render}{ggplot}(x, ...)

\method{md_render}{recordedplot}(x, value, dir = tempdir(), ...)

\method{md_render}{evaluate_evaluation}(x, ...)

\method{md_render}{data.frame}(x, ...)

\method{md_render}{gt_tbl}(x, value, dir = tempdir(), ...)

\method{md_render}{flextable}(x, ...)

md_server(board, update, session, parent, ...)

gen_md_ui(content = character())
}
\arguments{
\item{...}{Board attributes}

\item{document}{Initial document}

\item{value}{Current md block value}

\item{dir}{Where to place files (like images)}

\item{update, session, parent}{See \code{\link[blockr.ui:main]{blockr.ui::main_server()}}}

\item{content}{Initial content}
}
\description{
A markdown board pairs blockr functionality with a document builder.
}
