% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-ext.R, R/md-render.R
\name{new_md_extension}
\alias{new_md_extension}
\alias{md_render}
\alias{md_render.md_text}
\alias{md_render.file}
\alias{md_render.ggplot}
\alias{md_render.recordedplot}
\alias{md_render.evaluate_evaluation}
\alias{md_render.data.frame}
\alias{md_render.gt_tbl}
\alias{md_render.flextable}
\title{MD extension}
\usage{
new_md_extension(content = character(), ...)

md_render(x, value, dir = tempdir(), ...)

\method{md_render}{md_text}(x, value, dir = tempdir(), ...)

\method{md_render}{file}(x, value, dir = tempdir(), ...)

\method{md_render}{ggplot}(x, ...)

\method{md_render}{recordedplot}(x, value, dir = tempdir(), ..., .is_patchwork = FALSE)

\method{md_render}{evaluate_evaluation}(x, ...)

\method{md_render}{data.frame}(x, ...)

\method{md_render}{gt_tbl}(x, value, dir = tempdir(), ...)

\method{md_render}{flextable}(x, ...)
}
\arguments{
\item{content}{Initial document}

\item{...}{Forwarded to \code{\link[blockr.dock:extension]{blockr.dock::new_dock_extension()}}}

\item{x}{Plot}

\item{value}{Current md block value}

\item{dir}{Where to place files (like images)}

\item{.is_patchwork}{Is plot a patchwork plot?}
}
\description{
A markdown extension adds a document builder.
}
