% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.md-package}
\alias{blockr.md}
\alias{blockr.md-package}
\title{blockr.md: Markdown builder for blockr}
\description{
Provides a blockr.ui compatible markdown module.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
