% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{blockr_template}
\alias{blockr_template}
\title{Get path to bundled PowerPoint templates}
\usage{
blockr_template(template = NULL)
}
\arguments{
\item{template}{Name of the template file. If NULL, returns the templates directory.
Available templates: "bms-template.pptx" (default), "pandoc-default.pptx"}
}
\value{
Character path to the template file or templates directory
}
\description{
This function returns the path to PowerPoint templates bundled with the
blockr.md package. These templates are guaranteed to work with pandoc
without corruption issues.
}
\examples{
# Get BMS template (default)
bms_template <- blockr_template("bms-template.pptx")

# Get pandoc default template
default_template <- blockr_template("pandoc-default.pptx")

# List all available templates
list.files(blockr_template())

# Use with new_md_board
board <- new_md_board(pptx_template = blockr_template("bms-template.pptx"))

}
