% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{default_template}
\alias{default_template}
\title{Default template option}
\usage{
default_template()
}
\value{
Named character vector for the default template
}
\description{
Default template option
}
\examples{
default_template()

}
