% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-ext.R
\name{new_md_extension}
\alias{new_md_extension}
\title{Document extension}
\usage{
new_md_extension(content = character(), template = NULL, ...)
}
\arguments{
\item{content}{Document content passed as character vector}

\item{template}{(Optional) template for rendering via pandoc}

\item{...}{Forwarded to \code{\link[blockr.dock:extension]{blockr.dock::new_dock_extension()}}}
}
\description{
Create a markdown document, using blockr outputs.
}
