% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-board.R, R/md-ui.R
\name{new_md_board}
\alias{new_md_board}
\alias{is_md_board}
\alias{board_doc}
\alias{board_doc<-}
\alias{doc_ui}
\title{Markdown board}
\usage{
new_md_board(..., document = character())

is_md_board(x)

board_doc(x)

board_doc(x) <- value

doc_ui(id, x)
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{document}{Initial document}

\item{x}{Board object}

\item{value}{Replacement value}

\item{id}{Namespace ID}
}
\description{
A markdown board pairs blockr functionality with a document builder.
}
