% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-board.R, R/md-render.R, R/md-ui.R
\name{new_md_board}
\alias{new_md_board}
\alias{is_md_board}
\alias{board_doc}
\alias{board_doc<-}
\alias{md_render}
\alias{md_render.file}
\alias{md_render.ggplot}
\alias{md_render.recordedplot}
\alias{md_render.evaluate_evaluation}
\alias{md_render.data.frame}
\alias{md_render.gt_tbl}
\alias{doc_ui}
\title{Markdown board}
\usage{
new_md_board(..., document = character())

is_md_board(x)

board_doc(x)

board_doc(x) <- value

md_render(x, value, dir = tempdir(), ...)

\method{md_render}{file}(x, value, dir = tempdir(), ...)

\method{md_render}{ggplot}(x, value, dir = tempdir(), ...)

\method{md_render}{recordedplot}(x, value, dir = tempdir(), ...)

\method{md_render}{evaluate_evaluation}(x, value, dir = tempdir(), ...)

\method{md_render}{data.frame}(x, value, dir = tempdir(), ...)

\method{md_render}{gt_tbl}(x, value, dir = tempdir(), ...)

doc_ui(id, x)
}
\arguments{
\item{...}{Further (metadata) attributes}

\item{document}{Initial document}

\item{x}{Board object}

\item{value}{Current md block value}

\item{dir}{Where to place files (like images)}

\item{id}{Namespace ID}
}
\description{
A markdown board pairs blockr functionality with a document builder.
}
